### 2018-08-18 #901 @arfon
@sauln - please review the [submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html) for JOSS. Your project doesn't seem to have a paper.
