### 2018-08-18 #901 @sauln
Hi @arfon, sorry the paper was hiding in the Joss branch. I’ve merged it into master now.
