### 2018-08-18 #901 @arfon
@sauln - this PR should fix the paper build error: https://github.com/scikit-tda/ripser.py/pull/45
