### 2018-08-19 36280b0
Generated gh-pages for commit 8acb14a Author: Nathaniel Saul <nat@saulgill.com>
    fix docs appveyor badge
