### 2018-08-20 #901 @sauln
I think perspectives from outside the field of TDA would be beneficial. We hope that the library will be accessible to anyone with just an interest in TDA, so understanding how someone without previous background approaches the library would be helpful for us. That being said, a JOSS review might not be the best forum for that.
