### 2018-08-22 181cf55
updating documentation about license and relationship to ripser
