+++ docs/conf.py
@@ -19,14 +19,14 @@
 
 # -- Project information -----------------------------------------------------
 
-project = u'Ripser'
-copyright = u'2018, Christopher Tralie and Nathaniel Saul'
+project = u'Ripser.py'
+copyright = u'2018, (Documentation) Christopher Tralie and Nathaniel Saul'
 author = u'Christopher Tralie and Nathaniel Saul'
 
 # The short X.Y version
-version = u''
+version = u'0.2'
 # The full version, including alpha/beta/rc tags
-release = u'0.0.1'
+release = u'0.2.7'
 
 
 # -- General configuration ---------------------------------------------------
@@ -146,7 +146,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'ripser.tex', u'Ripser Documentation',
+    (master_doc, 'ripser.tex', u'Ripser.py Documentation',
      u'Christopher Tralie and Nathaniel Saul', 'manual'),
 ]
 
@@ -156,7 +156,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'ripser', u'Ripser Documentation',
+    (master_doc, 'ripser.py', u'Ripser.py Documentation',
      [author], 1)
 ]
 
@@ -167,9 +167,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'ripser', u'ripser Documentation',
-     author, 'Ripser', 'Persistence diagrams in Python.',
-     'Miscellaneous'),
+    (master_doc, 'ripser.py', u'Ripser.py Documentation',
+     author, 'Ripser.py', 'A Lean Persistent Homology Library for Python'),
 ]
 
 
