+++ docs/index.rst
@@ -9,15 +9,25 @@
 Ripser.py
 ==========
 
-Ripser.py is a renovated Python implementation of the Ripser package. 
+Ripser.py is a lean persistent homology package for Python. Building blazing fast C++ Ripser package as the core computational engine, Ripser.py provides an intuitive interface for 
 
-We provide a slim interface for computing persistence cohomology of sparse and dense data sets, visualizing persistence diagrams, computing lowerstar filtrations on images, and computing representative cochains. 
+- computing persistence cohomology of sparse and dense data sets, 
+- visualizing persistence diagrams, 
+- computing lowerstar filtrations on images, and 
+- computing representative cochains. 
+
+
+Additionally, through extensive testing and continuous integration, Ripser.py is easy to install on Mac, Linux, and Windows platforms.
 
 We supply a large set of interactive notebooks that demonstrate how to take advantage of all the features available.
 
-Through extensive testing and continuous integration, Ripser.py is easy to install on Mac, Linux, and Windows platforms.   
+You can find the source code on github at `Scikit-TDA/Ripser.py <https://github.com/scikit-tda/ripser.py>`_. For the original C++ library, see `Ripser/ripser <https://github.com/Ripser/ripser/releases/latest>`_.
 
-You can find the source code on github at `Scikit-TDA/ripser.py <https://github.com/scikit-tda/ripser.py>`_. For the original C++ library, see `Ripser/ripser <https://github.com/Ripser/ripser/releases/latest>`_.
+
+Example Usage
+--------------
+
+
 
 .. code:: python
 
