### 2018-08-22 4068425
Generated gh-pages for commit 39bd6c1 Merge: 181cf55 ae443d8 Author: Nathaniel Saul <nat@saulgill.com>
    Merge branch 'master' of https://github.com/scikit-tda/ripser.py
