### 2018-08-22 e11bdb9
Generated gh-pages for commit 0b88c1b Merge: 181cf55 ae443d8 Author: Nathaniel Saul <nat@saulgill.com>
    Merge branch 'master' of https://github.com/scikit-tda/ripser.py
