### 2018-08-30 #901 @arokem
👋 @lmcinnes: would you be willing to review this paper for JOSS?
