### 2018-08-30 #901 @arokem
👋 @maxrousseau: would you be willing to review this paper for JOSS?
