### 2018-08-31 e8e9863
Update paper.md
