### 2018-09-01 #901 @lmcinnes
@arokem , I would be willing to review this, but may need a little assistance, having never reviewed for JOSS before.
