### 2018-09-02 #901 @arokem
Thanks @lmcinnes! There is documentation that explains the principles and process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Let me know if you have questions. Don't be alarmed if it takes me a few days to get to your questions, though. For the time being, I will start the review, which will start another issue with a check-list for you to check the software against.
