### 2018-09-02 #925 @whedon
[REVIEW]: Ripser.py: A Lean Persistent Homology Library for Python
**Submitting author:** @sauln (<a href="http://orcid.org/0000-0002-8549-9810">Nathaniel Saul</a>)
**Repository:** <a href="https://github.com/scikit-tda/ripser.py" target ="_blank">https://github.com/scikit-tda/ripser.py</a>
**Version:** 0.2.4
