### 2018-09-02 #901 @arokem
@lmcinnes : the check list and more guidance should now be available for you on #925. I'm closing this one.
