### 2018-09-05 #925 @lmcinnes
First off, this is clearly a mature library with significant documentation available. It looks great!

I'm still working through the checklist, but some initial comments, all minor:

* the versions don't match, but I presume there were further patch releases after the initial submission?
* In the paper the reference for the original Ripser library isn't resolving; I think this is a capitalisation issue in the bibtex name.
* When testing the installation via pip there were errors in building the wheel (since pip hadn't installed numpy yet; it seemed to get around that and work, but the errors were disconcerting). Perhaps it would be best to add numpy as a dependency up front in the installation instructions?
* In contributions in the README it would be beneficial to cover how to report issues or questions.
* The checklist suggests everything that does have a DOI in the references should -- I believe many of the papers referenced have DOIs but the references don't list them. Perhaps @arokem can clarify this one?
