### 2018-09-05 32ed820
contributions blurb about issues and questions
