### 2018-09-05 16473b5
Add DOIs to many citations
