+++ paper.bib
@@ -1,7 +1,7 @@
 @article{bauer2017ripser,
   title={Ripser: a lean C++ code for the computation of Vietoris-Rips persistence barcodes},
   author={Bauer, Ulrich},
-  journal={Software available at https://github. com/Ripser/ripser},
+  journal={Software available at https://github.com/Ripser/ripser},
   year={2017}
 }
 
@@ -17,40 +17,56 @@
  year={2011}
 }
 
-
-@article{obayashi2018persistence,
-  title={Persistence diagrams with linear machine learning models},
-  author={Obayashi, Ippei and Hiraoka, Yasuaki and Kimura, Masao},
-  journal={Journal of Applied and Computational Topology},
-  volume={1},
-  number={3-4},
-  pages={421--449},
-  year={2018},
-  publisher={Springer}
+@Article{obayashi2018persistence,
+  author="Obayashi, Ippei
+  and Hiraoka, Yasuaki
+  and Kimura, Masao",
+  title="Persistence diagrams with linear machine learning models",
+  journal="Journal of Applied and Computational Topology",
+  year="2018",
+  month="Jun",
+  day="01",
+  volume="1",
+  number="3",
+  pages="421--449",
+  abstract="Persistence diagrams have been widely recognized as a compact descriptor for characterizing multiscale topological features in data. When many datasets are available, statistical features embedded in those persistence diagrams can be extracted by applying machine learnings. In particular, the ability for explicitly analyzing the inverse in the original data space from those statistical features of persistence diagrams is significantly important for practical applications. In this paper, we propose a unified method for the inverse analysis by combining linear machine learning models with persistence images. The method is applied to point clouds and cubical sets, showing the ability of the statistical inverse analysis and its advantages.",
+  issn="2367-1734",
+  doi="10.1007/s41468-018-0013-5",
+  url="https://doi.org/10.1007/s41468-018-0013-5"
 }
 
-
-
-@article{de2011persistent,
-  title={Persistent cohomology and circular coordinates},
-  author={De Silva, Vin and Morozov, Dmitriy and Vejdemo-Johansson, Mikael},
-  journal={Discrete \& Computational Geometry},
-  volume={45},
-  number={4},
-  pages={737--759},
-  year={2011},
-  publisher={Springer}
+@Article{de2011persistent,
+  author="de Silva, Vin
+  and Morozov, Dmitriy
+  and Vejdemo-Johansson, Mikael",
+  title="Persistent Cohomology and Circular Coordinates",
+  journal="Discrete {\&} Computational Geometry",
+  year="2011",
+  month="Jun",
+  day="01",
+  volume="45",
+  number="4",
+  pages="737--759",
+  abstract="Nonlinear dimensionality reduction (NLDR) algorithms such as Isomap, LLE, and Laplacian Eigenmaps address the problem of representing high-dimensional nonlinear data in terms of low-dimensional coordinates which represent the intrinsic structure of the data. This paradigm incorporates the assumption that real-valued coordinates provide a rich enough class of functions to represent the data faithfully and efficiently. On the other hand, there are simple structures which challenge this assumption: the circle, for example, is one-dimensional, but its faithful representation requires two real coordinates. In this work, we present a strategy for constructing circle-valued functions on a statistical data set. We develop a machinery of persistent cohomology to identify candidates for significant circle-structures in the data, and we use harmonic smoothing and integration to obtain the circle-valued coordinate functions themselves. We suggest that this enriched class of coordinate functions permits a precise NLDR analysis of a broader range of realistic data sets.",
+  issn="1432-0444",
+  doi="10.1007/s00454-011-9344-x",
[…]
