### 2018-09-07 #925 @lmcinnes
Hi @sauln,

It all looks good -- thanks for patching up the minor issues. My only remaining concern is the README provides example code that doesn't quite work. I think for the sklearn API version you want to remove the ``['dgm']``. Beyond that I think everything has been addressed.
