### 2018-09-07 8202b1a
fix fit_transform example and add image
