### 2018-09-07 #925 @sauln
@lmcinnes, thanks for catching that example issue. I've fixed the bug and added an example image for the output.
