### 2018-09-07 033f881
formatting paper.md
