### 2018-09-07 #925 @sauln
The paper looks good to me.  I did notice some of the references are rendering strangely, print the first name only some of the time. I'm not sure if this is intentional or if there is something wrong with our .bib.
