### 2018-09-10 c547353
update versions for finalizing JOSS review
