+++ docs/conf.py
@@ -24,9 +24,9 @@ copyright = u'2018, (Documentation) Christopher Tralie and Nathaniel Saul'
 author = u'Christopher Tralie and Nathaniel Saul'
 
 # The short X.Y version
-version = u'0.2'
+version = u'0.3'
 # The full version, including alpha/beta/rc tags
-release = u'0.2.7'
+release = u'0.3.0'
 
 
 # -- General configuration ---------------------------------------------------
