### 2018-09-10 #925 @lmcinnes
Thanks @arokem and @sauln! Glad to see this getting published.

On Sun, Sep 9, 2018 at 10:29 PM Ariel Rokem <notifications@github.com>
wrote:

> I have seen author first names in other JOSS papers as well, so I don't
> think this is anything unique to your paper.
>
> Regarding DOIs: I was also not able to find DOIs for these articles. I am
> surprised in particular that JMLR doesn't have DOIs (for the sklearn
> paper), but it does seem like it doesn't.
>
> One small comment: The paper currently suggests using pip install ripser
> (with a lower-case 'r'), while the README suggests using pip install
> Ripser (with a capital 'R'). Having just tried this, I realize that it
> doesn't matter which one you use, but I think that this might still confuse
> some readers of the article.
>
> Other than that - nice work! I think that the paper is ready to be
> accepted (thanks @lmcinnes <https://github.com/lmcinnes> for the
> review!). Once you have corrected this small item, please create an archive
> for the accepted version of the software (e.g., using Zenodo) and post it
> here.
>
