### 2018-09-10 ade904e
move version location to code instead of setup
