+++ ripser/__init__.py
@@ -1,5 +1,3 @@
 from .ripser import *
 
-# Enable access to version number
-import pkg_resources
-__version__ = pkg_resources.get_distribution('ripser').version
+from ._version import __version__
