+++ setup.py
@@ -14,6 +14,19 @@ except:
     sys.exit(1)
 
 
+import re
+VERSIONFILE="persim/_version.py"
+verstrline = open(VERSIONFILE, "rt").read()
+VSRE = r"^__version__ = ['\"]([^'\"]*)['\"]"
+mo = re.search(VSRE, verstrline, re.M)
+if mo:
+    verstr = mo.group(1)
+else:
+    raise RuntimeError("Unable to find version string in %s." % (VERSIONFILE,))
+
+
+
+
 with open('README.md') as f:
     long_description = f.read()
 
@@ -37,7 +50,7 @@ class CustomBuildExtCommand(build_ext):
 
 
 setup(name="ripser",
-      version='0.3.0',
+      version=verstr,
       description="A Lean Persistent Homology Library for Python",
       long_description=long_description,
       long_description_content_type="text/markdown",
