### 2018-09-10 58cd3e4
handle version in docs automatically
