### 2018-09-11 #925 @sauln
Thank you both @lmcinnes and @arokem for pushing this through. @ctralie and I are both excited to have this accepted.

I have incremented the version to 0.3.0, made a release, and uploaded it to zenodo. The badge is below (it appears there is a delay in the doi being registered, the link in the badge doesn't work just yet)

[![DOI](https://zenodo.org/badge/96122448.svg)](https://zenodo.org/badge/latestdoi/96122448)
