### 2018-09-14 d2494de
adding bibtex citation
