### 2018-10-04 c0c8117
update new ripser google tracking code
