### 2018-10-08 e10206f
Trying another compile flag
revert compiler flag to working state

add 3.7 to travis tests

attempt 3.7 travis work around.

Tweaking mac flags for debugging
