### 2018-10-08 c22955b
revert compiler flag to working state
