### 2018-10-08 780fb54
add 3.7 to travis tests
