+++ .travis.yml
@@ -4,6 +4,7 @@ python:
   - "3.4"
   - "3.5"
   - "3.6"
+  - "3.7"
 
 before_install:
     # Tricks to avoid matplotlib error about X11:
