### 2018-10-08 5ac5c75
trying to add Mac OS 10.7 lion to build tests
