+++ .travis.yml
@@ -7,10 +7,16 @@ python:
   
 matrix:
   include:
-    language: python
-    sudo: required
-    dist: xenial
-    python: "3.7"
+    - name: "Python 3.7 workaround"
+      language: python
+      sudo: required
+      dist: xenial
+      python: "3.7"
+    - name: "Mac 10.7, Very old Mac versions"
+      language: python
+      dist: lion
+      python: "3.6"
+    
   
 
 before_install:
