### 2018-10-08 f58756b
Special install commands for old MacOS versions (lion)
