### 2018-10-08 9f4ae7f
print platform data for debugging in travis
