### 2018-10-08 85403fc
using lion in travis ci
