+++ .travis.yml
@@ -14,6 +14,7 @@ matrix:
       python: "3.7"
     - name: "Mac 10.7, Very old Mac versions"
       language: python
+      os: osx
       dist: lion
       python: "3.6"
     
