### 2018-10-08 61e382c
Attempt explore old Mac versions on travis
