### 2018-10-11 349a937
trying to parse platform appropriately
