### 2018-10-11 1937dc5
install on travis with verbose flag
