### 2018-10-11 ccf69be
simply flag to be all macs
