### 2018-10-11 1ab3cf2
fixup! simply flag to be all macs
