+++ .travis.yml
@@ -22,7 +22,7 @@ before_install:
 
 install:
   - pip install Cython
-  - pip -v install .
+  - pip install .
   - pip install pytest-cov
 
 
