### 2018-10-15 dc9c50f
[WIP] Compiler fixes for Mac (#48)
Adding condition compile flag to accommodate old macs. Add 3.7 to travis tests
