### 2018-11-07 c91cdd9
updates to plotting defaults, accept ax argument (#53)
