+++ ripser/_version.py
@@ -1,1 +1,1 @@
-__version__ = "0.3.0"
+__version__ = "0.3.0.dev0"
