### 2018-11-14 1d352cc
Adding test that goes down the buggy path of execution with thresholds on H2
