### 2018-11-15 99cf0a5
Adding mex code to return persistent homology *and* cocycles
