+++ ripser/ripser.cpp
@@ -49,6 +49,12 @@ derivative works thereof, in binary and source code form.
 #include <sstream>
 #include <unordered_map>
 
+#ifdef MATLAB_MEX_FILE
+#include <mex.h>
+#define ASSEMBLE_REDUCTION_MATRIX
+#define USE_COEFFICIENTS
+#endif
+
 
 template <class Key, class T> class hash_map : public std::unordered_map<Key, T> {};
 typedef float value_t;
@@ -1170,3 +1176,115 @@ extern "C" {
     }
 }
 #endif
+
+
+
+
+
+
+
+
+#ifdef MATLAB_MEX_FILE
+//std::vector<value_t> distances, index_t n, value_t threshold, index_t dim_max, std::vector<std::vector<float>>& dgms
+void mexFunction(int nOutArray, mxArray *OutArray[], int nInArray, const mxArray *InArray[]) {
+	if (nInArray < 4) {
+		mexErrMsgTxt("Expecting D, threshold, modulus, dim_max");
+		return;
+	}
+	//Inputs: Distances, thresh, coeff, dim_max
+	const mwSize *dims;
+	dims = mxGetDimensions(InArray[0]);
+	size_t N = dims[0];
+	value_t* D = (value_t*)mxGetPr(InArray[0]);
+	coefficient_t modulus = (coefficient_t)(*((double*)mxGetPr(InArray[1])));
+	index_t dim_max = (index_t)(*((double*)mxGetPr(InArray[2])));
+	double threshold = *((double*)mxGetPr(InArray[3]));
+	value_t maxD = D[0];
+	for (size_t i = 1; i < N; i++) {
+		if (D[i] > maxD) {
+			maxD = D[i];
+		}
+	}
+	//Check to make sure there are enough output arguments to hold
+	//all persistence diagrams
+	if (nOutArray < 1) {
+		mexErrMsgTxt("Need at least one output to hold cell array of persistence diagrams");
+		return;
+	}
+	bool do_cocycles = false;
+	//Setup cell array to hold output
+	mwSize cdims[1];
+	cdims[0] = dim_max+1;
+	mxArray* PDs = mxCreateCellArray(1, cdims);
+	OutArray[0] = PDs;
+	mxArray* cocycles;
+	if (nOutArray == 2) {
+		cocycles = mxCreateCellArray(1, cdims);
+		OutArray[1] = cocycles;
+		do_cocycles = true;
+	}
+	//Setup distance matrix, coefficient tables, etc
+	std::vector<value_t> distances(D, D+N);
+	compressed_lower_distance_matrix dist = compressed_lower_distance_matrix(compressed_upper_distance_matrix(std::move(distances)));
+	float ratio = 1.0; //TODO: This seems like a dummy parameter at the moment
+	value_t min = std::numeric_limits<value_t>::infinity(),
+	        max = -std::numeric_limits<value_t>::infinity(), max_finite = max;
+	int num_edges = 0;
+	value_t enclosing_radius = std::numeric_limits<value_t>::infinity();
+	for (size_t i = 0; i < dist.size(); ++i) {
+		value_t r_i = -std::numeric_limits<value_t>::infinity();
+		for (size_t j = 0; j < dist.size(); ++j) r_i = std::max(r_i, dist(i, j));
+		enclosing_radius = std::min(enclosing_radius, r_i);
+	}
+	if (threshold == std::numeric_limits<value_t>::max()) threshold = enclosing_radius;
+	for (auto d : dist.distances) {
+		min = std::min(min, d);
+		max = std::max(max, d);
[…]
