### 2018-11-15 cc3b4c1
Adding basic plotting utilities
