### 2018-11-16 50309a4
Adding sparse matrix support and lower star filtrations, and coming up with a better API with string named parsed arguments.  Also adding a notebook demonstrating all of the features
