### 2018-11-16 c0dfd3f
Adding a terse readme
