+++ ripser/README.md
@@ -1,0 +1,3 @@
+# Matlab wrapper for ripser
+
+All of the code and mex files (for Ubuntu, Windows, and Mac) to wrap around ripser are in this directory.  Have a look at the "MatlabBasics" notebook for usage.
