### 2019-01-24 95771a0
travis tests in conda and macosx
