+++ setup.py
@@ -1,5 +1,6 @@
 import sys
 import os
+import platform
 
 from setuptools import setup
 from setuptools.extension import Extension
@@ -25,15 +26,8 @@ else:
     raise RuntimeError("Unable to find version string in %s." % (VERSIONFILE,))
 
 
-
-
 with open('README.md') as f:
     long_description = f.read()
-
-options = ["-std=c++11", "-Ofast", "-D_hypot=hypot"]
-
-if sys.version_info[0] == 2:
-    options.append("-fpermissive")
 
 
 class CustomBuildExtCommand(build_ext):
@@ -49,30 +43,61 @@ class CustomBuildExtCommand(build_ext):
         build_ext.run(self)
 
 
-setup(name="ripser",
-      version=verstr,
-      description="A Lean Persistent Homology Library for Python",
-      long_description=long_description,
-      long_description_content_type="text/markdown",
-      author="Chris Tralie, Nathaniel Saul",
-      author_email="chris.tralie@gmail.com, nathaniel.saul@wsu.edu",
-      url="https://ripser.scikit-tda.org",
-      license='MIT',
-      packages=['ripser'],
-      ext_modules=cythonize(Extension("pyRipser",
-                                      sources=["ripser/pyRipser.pyx"],
-                                      define_macros=[("USE_COEFFICIENTS", 1),
-                                                     ("NDEBUG", 1), ("ASSEMBLE_REDUCTION_MATRIX", 1)],
-                                      extra_compile_args=options,
-                                      language="c++"
-                                      )),
+extra_compile_args = ["-Ofast", "-D_hypot=hypot"]
+extra_link_args = []
 
-      install_requires=[
-          'Cython',
-          'numpy',
-          'scipy',
-          'matplotlib',
-          'scikit-learn'
-      ],
-      cmdclass={'build_ext': CustomBuildExtCommand},
-      )
+if platform.system() == "Windows":
+    extra_compile_args.extend([
+        "/std:c++latest", 
+        "/EHsc"
+    ])
+elif platform.system() == "Darwin":
+    extra_compile_args.extend([
+        '-std=c++11', 
+        "-mmacosx-version-min=10.9"
+    ])
+    extra_link_args.extend([
+        "-stdlib=libc++", 
+        "-mmacosx-version-min=10.9"
+    ])
+else:
+    extra_compile_args.extend([
+        "-std=c++11"
+    ])
+
+
+ext_modules = Extension(
+    "pyRipser",
+    sources=["ripser/pyRipser.pyx"],
+    define_macros=[
+        ("USE_COEFFICIENTS", 1),
+        ("NDEBUG", 1), 
+        ("ASSEMBLE_REDUCTION_MATRIX", 1)
+    ],
+    extra_compile_args=extra_compile_args,
+    extra_link_args=extra_link_args,
+    language="c++"
+)
+
+
+setup(
+    name="ripser",
+    version=verstr,
+    description="A Lean Persistent Homology Library for Python",
+    long_description=long_description,
+    long_description_content_type="text/markdown",
+    author="Chris Tralie, Nathaniel Saul",
+    author_email="chris.tralie@gmail.com, nathaniel.saul@wsu.edu",
+    url="https://ripser.scikit-tda.org",
+    license='MIT',
+    packages=['ripser'],
+    ext_modules=cythonize(ext_modules),
+    install_requires=[
+        'Cython',
+        'numpy',
+        'scipy',
+        'matplotlib',
+        'scikit-learn'
+    ],
+    cmdclass={'build_ext': CustomBuildExtCommand},
+)
