### 2019-01-25 578d66a
Travis tests for condo and macOS (#60)
* Update README.md

* adding bibtex citation

* Update README.md

* travis tests in conda and macosx
