+++ .travis.yml
@@ -1,30 +1,44 @@
 language: python
-python:
-  - "2.7"
-  - "3.4"
-  - "3.5"
-  - "3.6"
-  
+
 matrix:
   include:
-    - name: "Python 3.7 workaround"
-      language: python
-      sudo: required
+    - python: 3.6
+    - python: 3.7
       dist: xenial
-      python: "3.7"
-    
+      sudo: true
+    - env: DISTRIB="conda" PYTHON_VERSION="3.6"
+    - env: DISTRIB="conda" PYTHON_VERSION="3.7"
+    - os: osx
+      language: generic
+      env: DISTRIB="conda" PYTHON_VERSION="3.6"
+    - os: osx
+      language: generic
+      env: DISTRIB="conda" PYTHON_VERSION="3.7"
+    - os: osx
+      language: generic
+      env: PYTHON_VERSION="3.6"
+    - os: osx
+      language: generic
+      env: PYTHON_VERSION="3.7"
+
 before_install:
     # Tricks to avoid matplotlib error about X11:
     # 'no display name and no $DISPLAY environment variable'
-    # http://docs.travis-ci.com/user/gui-and-headless-browsers/#Starting-a-Web-Server
-    - "export DISPLAY=:99.0"
-    - "sh -e /etc/init.d/xvfb start"
 
-install:
-  - pip install Cython
-  - pip install .
-  - pip install pytest-cov
+    # enable gui, see https://docs.travis-ci.com/user/gui-and-headless-browsers
+    - if [ "$TRAVIS_OS_NAME" == "linux" ]; then
+            export DISPLAY=:99.0;
+            sh -e /etc/init.d/xvfb start;
+            sleep 3;
+      fi
+    - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew list python &>/dev/null || brew install python; fi
+    - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew list python3 &>/dev/null || brew install python3; fi
+    - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew install pyenv-virtualenv; fi
 
+services:
+  - xvfb
+
+install: source ci_scripts/install.sh
 
 # Installing matplotlib is long processes. Using a cache allows
 # to fasten the tests by reducing drastically the install time.
