### 2019-02-04 d9b5c4c
Adding patch for the case where dense matrices have nonzero entries on the diagonal
