### 2019-02-04 fab7957
Adding greedy permutation to the API, along with examples
