+++ ripser/ripser.py
@@ -37,6 +37,82 @@ from pyRipser import doRipsFiltrationDM as DRFDM
 from pyRipser import doRipsFiltrationDMSparse as DRFDMSparse
 
 
+def get_greedy_perm_pc(X, N = -1, metric="euclidean"):
+    """
+    Compute a furthest point sampling permutation of a set
+    of points
+    Parameters
+    ----------
+    X: ndarray(N, d)
+        Point cloud with N points in d dimensions
+    N: int
+        Number of points to take in the permutation
+    metric: string or callable
+        The metric to use when calculating distance between instances in a 
+        feature array. If metric is a string, it must be one of the options 
+        specified in PAIRED_DISTANCES, including "euclidean", "manhattan", 
+        or "cosine". Alternatively, if metric is a callable function, it is 
+        called on each pair of instances (rows) and the resulting value 
+        recorded. The callable should take two arrays from X as input and 
+        return a value indicating the distance between them.
+    Returns
+    -------
+    perm: ndarray(N)
+        Indices of points in the greedy permutation
+    lambdas: ndarray(N)
+        Covering radii at different points
+    """
+    if N == -1:
+        N = X.shape[0]
+    #By default, takes the first point in the list to be the
+    #first point in the permutation, but could be random
+    perm = np.zeros(N, dtype=np.int64)
+    lambdas = np.zeros(N)
+    ds = pairwise_distances(X, X[0, :][None, :], metric=metric)
+    for i in range(1, N):
+        idx = np.argmax(ds)
+        perm[i] = idx
+        lambdas[i-1] = ds[idx]
+        ds = np.minimum(ds, pairwise_distances(X, X[idx, :][None, :], metric=metric))
+    lambdas[-1] = np.max(ds)
+    return (perm, lambdas)
+
+
+
+def get_greedy_perm_dm(D, N = -1):
+    """
+    Compute a furthest point sampling permutation of a set
+    of points
+    Parameters
+    ----------
+    D: ndarray(N, N)
+        NxN distance matrix
+    N: int
+        Number of points to take in the permutation
+    Returns
+    -------
+    perm: ndarray(N)
+        Indices of points in the greedy permutation
+    lambdas: ndarray(N)
+        Covering radii at different points
+    """
+    if N == -1:
+        N = D.shape[0]
+    #By default, takes the first point in the list to be the
+    #first point in the permutation, but could be random
+    perm = np.zeros(N, dtype=np.int64)
+    lambdas = np.zeros(N)
+    ds = D[0, :]
+    for i in range(1, N):
+        idx = np.argmax(ds)
+        perm[i] = idx
+        lambdas[i-1] = ds[idx]
+        ds = np.minimum(ds, D[idx, :])
+    lambdas[-1] = np.max(ds)
+    return (perm, lambdas)
+
+
+
 def ripser(
     X,
     maxdim=1,
@@ -45,6 +121,7 @@ def ripser(
     distance_matrix=False,
     do_cocycles=False,
     metric="euclidean",
+    n_perm = -1,
 ):
     """Compute persistence diagrams for X data array. If X is not a distance matrix, it will be converted to a distance matrix using the chosen metric.
 
@@ -82,6 +159,14 @@ def ripser(
         called on each pair of instances (rows) and the resulting value 
         recorded. The callable should take two arrays from X as input and 
         return a value indicating the distance between them.
+    
+    n_perm: int
+        The number of points to subsample in a "greedy permutation,"
+        or a furthest point sampling of the points.  These points
+        will be used in lieu of the full point cloud for a faster
+        computation, at the expense of some accuracy, which can 
+        be bounded as a maximum bottleneck distance to all diagrams
+        on the original point set
 
     Returns
     -------
@@ -92,13 +177,25 @@ def ripser(
[…]
