### 2019-02-04 214f47e
Adding a couple of more exceptions and tests
