### 2019-02-05 e607cc3
Making tests slightly faster
