### 2019-02-13 c54e379
update docs to new style
