+++ docs/reference/stubs/ripser.ripser.rst
@@ -1,0 +1,6 @@
+ripser.ripser
+=============
+
+.. currentmodule:: ripser
+
+.. autofunction:: ripser
