+++ docs/reference/index.rst
@@ -1,0 +1,11 @@
+API Reference
+--------------
+
+.. autosummary::
+    :toctree: stubs
+    :nosignatures:
+
+    ripser.ripser
+    ripser.plot_dgms
+    ripser.Rips
+
