### 2019-02-13 ca612db
Generated gh-pages for commit 21b0937 Merge: c54e379 5497f25 Author: Nathaniel Saul <nat@saulgill.com>
    Merge branch 'master' of https://github.com/scikit-tda/ripser.py
