+++ _sources/reference/stubs/ripser.plot_dgms.rst.txt
@@ -1,0 +1,6 @@
+ripser.plot\_dgms
+=================
+
+.. currentmodule:: ripser
+
+.. autofunction:: plot_dgms
