+++ _sources/reference/stubs/ripser.Rips.rst.txt
@@ -1,0 +1,25 @@
+ripser.Rips
+===========
+
+.. currentmodule:: ripser
+
+.. autoclass:: Rips
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~Rips.__init__
+      ~Rips.fit_transform
+      ~Rips.plot
+      ~Rips.transform
+   
+   
+
+   
+   
+   
