### 2019-02-14 44f5f9a
update all notebooks to use persim and tadasets, all examples run
