### 2019-02-14 35e5308
Generated gh-pages for commit 44f5f9a Author: Nathaniel Saul <nat@saulgill.com>
    update all notebooks to use persim and tadasets, all examples run
