+++ _sources/notebooks/Lower Star Time Series.ipynb.txt
@@ -17,11 +17,13 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "from ripser import ripser, plot_dgms\n",
     "import numpy as np\n",
     "import matplotlib.pyplot as plt\n",
     "from scipy import sparse\n",
-    "import time"
+    "import time\n",
+    "\n",
+    "from ripser import ripser\n",
+    "from persim import plot_diagrams"
    ]
   },
   {
@@ -39,7 +41,7 @@
     {
      "data": {
       "text/plain": [
-       "Text(0.5,0,'t')"
+       "Text(0.5, 0, 't')"
       ]
      },
      "execution_count": 2,
@@ -48,12 +50,14 @@
     },
     {
      "data": {
-      "image/png": "iVBORw0KGgoAAAANSUhEUgAAAW4AAAEYCAYAAAB82RxTAAAABHNCSVQICAgIfAhkiAAAAAlwSFlzAAALEgAACxIB0t1+/AAAADl0RVh0U29mdHdhcmUAbWF0cGxvdGxpYiB2ZXJzaW9uIDIuMS4yLCBodHRwOi8vbWF0cGxvdGxpYi5vcmcvNQv5yAAAIABJREFUeJzt3Xd4VNe57/Hvqy4kgRAqgCQQCBDNVFFMBwPGvcQ2Ji5JTnIc+6Q5zSn35F4nOU45yUlxcuOEGzu244I7sbGNwTQZTBMdJCQESKhr1FBvM+v+IeEQLNCMNHuKeD/Po8cqe9Z+5wH9vFh7FTHGoJRSyn8EeLsApZRSrtHgVkopP6PBrZRSfkaDWyml/IwGt1JK+RkNbqWU8jMa3Eop5Wc0uJVSys9ocCu/JiI/F5FHLWp7n4hMsqJtpfpCg1v5LRGJAx4E/iIioSLytIgUiEi9iBwSkRtcbK9YRKZd9K1fAz9xY72Xtq9Ur2hwK3/2eeA9Y0wzEAQUAouBQcCPgFdFJMWZhkQkFogHsi/69tvAUhEZ1sNrHxeRx3vRvlK9osGtPEZEkkXkTRGxiUiViPyx6/sTRGS7iNSKyAkRufWi13yvq6daLyI5InLdRU3eAOwAMMY0GmMeN8bkG2McxpgNwFlgZlc7oSJSJyJ2EWno+rCLSIuIfInO0A8AqrpqCzLGtAAHgJV9fN9jumu/L22qq5v+5VEeISKBwAZgK/AAYAfSRSQYeAd4hs6AXAD8Q0TSu176VWCWMaakq/cceFGz1wA5l7lfAjAOOAFgjGkVkW8Aa4wxK7uuyQW+ZIzJEJFQYJExZvUlTWUDU/vy3o0xeSLyncu0r5TLtMetPGU2MBz4blfvuMUYsxOYC0QCvzDGtBljttIZ8GvoDPdQYKKIBHf1pk9f1GY0UH/pjbr+Z/Ai8Jwx5uRFP7
[…]
