+++ _sources/notebooks/Sparse Distance Matrices.ipynb.txt
@@ -17,12 +17,15 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "from ripser import ripser, plot_dgms\n",
     "import numpy as np\n",
     "import matplotlib.pyplot as plt\n",
     "from sklearn import datasets\n",
     "from sklearn.metrics.pairwise import pairwise_distances\n",
-    "from scipy import sparse"
+    "from scipy import sparse\n",
+    "\n",
+    "from ripser import ripser\n",
+    "from persim import plot_diagrams\n",
+    "import tadasets"
    ]
   },
   {
@@ -61,17 +64,53 @@
    "metadata": {},
    "outputs": [
     {
+     "data": {
+      "text/plain": [
+       "<matplotlib.collections.PathCollection at 0x11f968320>"
+      ]
+     },
+     "execution_count": 3,
+     "metadata": {},
+     "output_type": "execute_result"
+    },
+    {
+     "data": {
[…]
