+++ docs/reference/stubs/ripser.plot_dgms.rst
@@ -1,6 +1,0 @@
-ripser.plot\_dgms
-=================
-
-.. currentmodule:: ripser
-
-.. autofunction:: plot_dgms
