### 2019-02-18 8a6e4ef
Addressing all aspects of the code review except for linting with Black
