+++ ripser/ripser.py
@@ -37,79 +37,70 @@ from pyRipser import doRipsFiltrationDM as DRFDM
 from pyRipser import doRipsFiltrationDMSparse as DRFDMSparse
 
 
-def get_greedy_perm_pc(X, N = -1, metric="euclidean"):
+def dpoint2pointcloud(X, i, metric):
     """
-    Compute a furthest point sampling permutation of a set
-    of points
+    Return the distance from the ith point in a Euclidean point cloud
+    to the rest of the points
     Parameters
     ----------
-    X: ndarray(N, d)
-        Point cloud with N points in d dimensions
-    N: int
+    X: ndarray (n_samples, n_features)
+        A numpy array of data 
+    i: int
+        The index of the point from which to return all distances
+    metric: string or callable
+        The metric to use when calculating distance between instances in a 
+        feature array
+    """
+    ds = pairwise_distances(X, X[i, :][None, :], metric=metric).flatten()
+    ds[i] = 0
+    return ds
+
+def get_greedy_perm(X, n_perm = None, distance_matrix=False, metric="euclidean"):
+    """
+    Compute a furthest point sampling permutation of a set of points
+    Parameters
+    ----------
+    X: ndarray (n_samples, n_features)
+        A numpy array of either data or distance matrix
+    distance_matrix: bool
+        Indicator that X is a distance matrix, if not we compute 
+        distances in X using the chosen metric.
+    n_perm: int
         Number of points to take in the permutation
     metric: string or callable
         The metric to use when calculating distance between instances in a 
-        feature array. If metric is a string, it must be one of the options 
-        specified in PAIRED_DISTANCES, including "euclidean", "manhattan", 
-        or "cosine". Alternatively, if metric is a callable function, it is 
-        called on each pair of instances (rows) and the resulting value 
-        recorded. The callable should take two arrays from X as input and 
-        return a value indicating the distance between them.
+        feature array
     Returns
     -------
-    perm: ndarray(N)
+    idx_perm: ndarray(n_perm)
         Indices of points in the greedy permutation
-    lambdas: ndarray(N)
+    lambdas: ndarray(n_perm)
         Covering radii at different points
+    dperm2all: ndarray(n_perm, n_samples)
+        Distances from points in the greedy permutation to points
+        in the original point set
     """
-    if N == -1:
-        N = X.shape[0]
+    if not n_perm:
+        n_perm = X.shape[0]
     #By default, takes the first point in the list to be the
     #first point in the permutation, but could be random
-    perm = np.zeros(N, dtype=np.int64)
-    lambdas = np.zeros(N)
-    ds = pairwise_distances(X, X[0, :][None, :], metric=metric)
-    for i in range(1, N):
+    idx_perm = np.zeros(n_perm, dtype=np.int64)
+    lambdas = np.zeros(n_perm)
+    if distance_matrix:
+        dpoint2all = lambda i: X[i, :]
+    else:
+        dpoint2all = lambda i: dpoint2pointcloud(X, i, metric)
+    ds = dpoint2all(0)
+    dperm2all = [ds]
+    for i in range(1, n_perm):
         idx = np.argmax(ds)
-        perm[i] = idx
+        idx_perm[i] = idx
         lambdas[i-1] = ds[idx]
-        ds = np.minimum(ds, pairwise_distances(X, X[idx, :][None, :], metric=metric))
+        dperm2all.append(dpoint2all(idx))
+        ds = np.minimum(ds, dperm2all[-1])
     lambdas[-1] = np.max(ds)
-    return (perm, lambdas)
-
-
-
-def get_greedy_perm_dm(D, N = -1):
-    """
-    Compute a furthest point sampling permutation of a set
-    of points
-    Parameters
-    ----------
-    D: ndarray(N, N)
[…]
