+++ test/test_ripser.py
@@ -13,7 +13,7 @@ def makeSparseDM(X, thresh):
     Helper function to make a sparse distance matrix
     """
     N = X.shape[0]
-    D = pairwise_distances(X, metric='euclidean')
+    D = pairwise_distances(X, metric="euclidean")
     [I, J] = np.meshgrid(np.arange(N), np.arange(N))
     I = I[D <= thresh]
     J = J[D <= thresh]
@@ -21,84 +21,89 @@ def makeSparseDM(X, thresh):
     return sparse.coo_matrix((V, (I, J)), shape=(N, N)).tocsr()
 
 
-class TestLibrary():
+class TestLibrary:
     # Does the library install in scope? Are the objects in scope?
     def test_import(self):
         import ripser
         from ripser import ripser, plot_dgms
+
         assert 1
 
 
-class TestTransform():
+class TestTransform:
     def test_input_warnings(self):
         data = np.random.random((3, 10))
 
-        with pytest.warns(UserWarning, match='has more columns than rows') as w:
+        with pytest.warns(UserWarning, match="has more columns than rows") as w:
             ripser(data)
 
         data = np.random.random((3, 3))
-        with pytest.warns(UserWarning, match='input matrix is square, but the distance_matrix') as w:
+        with pytest.warns(
+            UserWarning, match="input matrix is square, but the distance_matrix"
+        ) as w:
             ripser(data)
 
     def test_non_square_dist_matrix(self):
         data = np.random.random((3, 10))
 
-        with pytest.raises(Exception, match='Distance matrix is not square'):
+        with pytest.raises(Exception, match="Distance matrix is not square"):
             ripser(data, distance_matrix=True)
-    
+
     def test_too_many_perm(self):
         X = np.zeros((10, 2))
 
-        with pytest.raises(Exception, match='points in greedy permutation is greater'):
-            ripser(X, n_perm = 11)
-    
+        with pytest.raises(Exception, match="points in greedy permutation is greater"):
+            ripser(X, n_perm=11)
+
     def test_too_few_perm(self):
         X = np.zeros((10, 2))
 
-        with pytest.raises(Exception, match='Should be a strictly positive number'):
-            ripser(X, n_perm = -1)
+        with pytest.raises(Exception, match="Should be a strictly positive number"):
+            ripser(X, n_perm=-1)
 
     def test_sparse_greedyperm(self):
         D = np.zeros((4, 4))
-        
+
         with pytest.raises(Exception):
-            ripser(sparse.coo_matrix(D), distance_matrix=True, n_perm = 2)
+            ripser(sparse.coo_matrix(D), distance_matrix=True, n_perm=2)
 
 
-class TestParams():
+class TestParams:
     def test_defaults(self):
         data = np.random.random((100, 3))
-        dgms = ripser(data)['dgms']
+        dgms = ripser(data)["dgms"]
         assert len(dgms) == 2
 
     def test_coeff(self):
         np.random.seed(3100)
         data = np.random.random((100, 3))
 
-        dgm3 = ripser(data, coeff=3)['dgms']
-        dgm2 = ripser(data)['dgms']
-        assert dgm2 is not dgm3, "This is a vacuous assertion, we only care that the above operations did not throw errors"
+        dgm3 = ripser(data, coeff=3)["dgms"]
[…]
