+++ test/test_ripser_sklearn.py
@@ -12,7 +12,7 @@ def makeSparseDM(X, thresh):
     Helper function to make a sparse distance matrix
     """
     N = X.shape[0]
-    D = pairwise_distances(X, metric='euclidean')
+    D = pairwise_distances(X, metric="euclidean")
     [I, J] = np.meshgrid(np.arange(N), np.arange(N))
     I = I[D <= thresh]
     J = J[D <= thresh]
@@ -20,11 +20,12 @@ def makeSparseDM(X, thresh):
     return sparse.coo_matrix((V, (I, J)), shape=(N, N)).tocsr()
 
 
-class TestLibrary():
+class TestLibrary:
     # Does the library install in scope? Are the objects in scope?
     def test_import(self):
         import ripser
         from ripser import Rips
+
         assert 1
 
     def test_instantiate(self):
@@ -32,21 +33,23 @@ class TestLibrary():
         assert rip is not None
 
 
-class TestTransform():
+class TestTransform:
     def test_input_warnings(self):
 
         rips = Rips()
         data = np.random.random((3, 10))
 
-        with pytest.warns(UserWarning, match='has more columns than rows') as w:
+        with pytest.warns(UserWarning, match="has more columns than rows") as w:
             rips.transform(data)
 
         data = np.random.random((3, 3))
-        with pytest.warns(UserWarning, match='input matrix is square, but the distance_matrix') as w:
+        with pytest.warns(
+            UserWarning, match="input matrix is square, but the distance_matrix"
+        ) as w:
             rips.transform(data)
 
     def test_non_square_dist_matrix(self):
-        
+
         rips = Rips()
         data = np.random.random((3, 10))
 
@@ -54,21 +57,21 @@ class TestTransform():
             rips.transform(data, distance_matrix=True)
 
 
-class TestParams():
+class TestParams:
     def test_verbose_true(self):
         data = np.array([[0.0, 0.0], [10.0, 0.0], [10.0, 10.0], [0.0, 10.0]])
         rips = Rips(verbose=True)
         dgm = rips.fit_transform(data)
         assert len(dgm) == 2
-        assert dgm[0].shape == (4,2)
-        assert dgm[1].shape == (1,2)
+        assert dgm[0].shape == (4, 2)
+        assert dgm[1].shape == (1, 2)
 
     def test_verbose_false(self):
         data = np.array([[0.0, 0.0], [10.0, 0.0], [10.0, 10.0], [0.0, 10.0]])
         rips = Rips(verbose=False)
         dgm = rips.fit_transform(data)
-        assert dgm[0].shape == (4,2)
-        assert dgm[1].shape == (1,2)
+        assert dgm[0].shape == (4, 2)
+        assert dgm[1].shape == (1, 2)
 
     def test_defaults(self):
         data = np.random.random((100, 3))
@@ -88,7 +91,9 @@ class TestParams():
 
         rips2 = Rips(coeff=2)
         dgm2 = rips2.fit_transform(data)
-        assert dgm2 is not dgm3, "This is a vacuous assertion, we only care that the above operations did not throw errors"
+        assert (
+            dgm2 is not dgm3
+        ), "This is a vacuous assertion, we only care that the above operations did not throw errors"
 
     def test_maxdim(self):
         np.random.seed(3100)
@@ -126,8 +131,10 @@ class TestParams():
         thresh = 1.1
 
         # Do dense filtration with threshold
-        data = datasets.make_circles(n_samples=100)[
-            0] + 5 * datasets.make_circles(n_samples=100)[0]
+        data = (
[…]

[…]
