### 2019-02-19 98dc513
Generated gh-pages for commit b749d0b Merge: 73ce108 1163b9d Author: Nathaniel Saul <nat@saulgill.com>
    Merge pull request #63 from scikit-tda/greedyperm
