+++ _sources/notebooks/Greedy Subsampling for Fast Approximate Computation.ipynb.txt
@@ -1,0 +1,234 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Greedy Permutation\n",
+    "\n",
+    "Sometimes, you will want to perform persistent homology on a point cloud with many points, but as you may have discovered, ripser.py grinds to a halt even on H1 if the number of points exceeds around 1000.  To mitigate this, we provide a routine that intelligently subsamples the points via \"furthest point sampling.\"  Please see\n",
+    "https://gist.github.com/ctralie/128cc07da67f1d2e10ea470ee2d23fe8\n",
+    "or refer to the paper\n",
+    "\n",
+    "\n",
+    "[1] Nicholas J Cavanna, Mahmoodreza Jahanseir, Donald R Sheehy, \"A geometric perspective on sparse filtrations\" \n",
+    "\n",
+    "\n",
+    "for more details.\n",
+    "\n"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 13,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "The autoreload extension is already loaded. To reload it, use:\n",
+      "  %reload_ext autoreload\n"
+     ]
+    }
+   ],
+   "source": [
+    "%load_ext autoreload\n",
+    "%autoreload 2\n",
+    "import time\n",
+    "import numpy as np\n",
+    "import matplotlib.pyplot as plt\n",
+    "from sklearn import datasets\n",
+    "\n",
+    "import tadasets\n",
+    "from persim import plot_diagrams, bottleneck\n",
+    "from persim.plot import bottleneck_matching\n",
+    "from ripser import ripser\n"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 6,
+   "metadata": {},
+   "outputs": [
[…]
