### 2019-02-19 8571e69
Update readme - how plotting is handled
