### 2019-02-28 bcfa209
Generated gh-pages for commit 407c1bc Merge: 024aec8 8571e69 Author: Nathaniel Saul <nat@saulgill.com>
    Merge branch 'master' of https://github.com/scikit-tda/ripser.py
